Fix default duration of simulation scenarios
The previous version of the code had the request generator function
set with an extremely low scenarios duration that was used for
debugging puroposes but could not be used for reaistic simulations.
As a result, it was not possible to properly reproduce the results
of the ACM SIGCOMM ICN'13 hash-routing paper without previously
increasing the duration of the simulations.

This commit reset the default simulation duration to the one used for
generating the results of the paper.