Add EXPERIMENT_QUEUE setting
This commit adds the EXPERIMENT_QUEUE setting which allows users to
select more accurately the combinations of parameters that will be
tested.

Without this setting, the orchestrator runs simulations
for each possible combination of experiment parameters (topology,
strategy, alpha, cache size). This settings, instead, allows users
to specify exactly the list of combinations of parameters to test.