Fix minor issue with HR Hybrid AM strategy
This commit fixes a bug with the HR Hybrid AM strategy that caused
a miscalculation of the path stretch threshold. As a result, the path
stretch treshold used was lower than the one specified.

This bug was introduced in version 0.2.0. It was not present in
version 0.1.x and did not affect the results of the ACM SIGCOMM
ICN'13 hash-routing paper.