Add function for formatting experiment durations and ETA
This commit adds a generic function to the util module that takes
as input a quantity of seconds and returns a nicely formatted
string representing the time, i.e. 90 -> 1m 30s.

This function is now used only to log the duration of an experiment
and the ETA for the entire experiment set.

This commit also adds test cases for this function.