Modify set_rand_insert to copy cache and return modified
Before, the set_rand_insert function of the cache module modified a
caching policy passed as argument to make it perform random item
insertion without returning anything. With this modification, the
instance passed as argument is not modified. The modified cache
instance is returned by the function.

This commit also renames the function to rand_insert_cache