Add failure messages in parallel execution
It makes Icarus report failures in the execution of an experiment
even when running in a process pool. Before, any issue would not
be shown to the user, who could know of a failure only by looking
at the result set.

In addition to that, it also prints in summary reports the
number of failed experiments.