Add Segmented LRU (SLRU) cache policy implementation
This policy divides the cache space into a number of segments of
equal size each operating according to an LRU policy. When a new
item isinserted to the cache, it is placed on the top entry of the
bottom segment. Each subsequent hit promotes the item to the top
entry of the segment above. When an item is evicted from a segment,
it is demoted to the top entry of the segment immediately below.
An item is evicted from the cache when it is evicted from the
bottom segment.