Greatly improve performance when using large content catalogues
This commit removes two bottlenecks affecting performance when large
catalogues are used. The first bottleneck was a search in a list that
was executed at each cache miss. This has now been replaced by a search
in a set. The second bottleneck was a double unnecessary deepcopy of the
network topology.