Move content placement at the end of orchestration
This move is required because if the workload contains many content
objects, the content placement considerably increases the size of the
topology. Therefore any operation requiring topology deep copy is going
to take a noticeably longer time. Hence content placement is moved to
the end, so that other operations can be done earlier on in the process.