Add Belady's MIN implementation
Belady's MIN policy is the provably optimal cache replacement policy
under arbitrary workload. Each time an item is inserted into a full cache,
it evicts the item that will be requested next the latest.

This policy is not implementable in practice because it requires knowledge
of future requests, however it is very useful as a theoretical performance
upper bound.