Remove Python version checks in test modules
The test cases require features of the unittest package that are
available only from Python 2.7 and backported to earlier versions of
Python in the unittest2 package.

At the beginning of each test module, there was a check that, in case of
Python version < 2.7 tried to import unittest2. Since Icarus does not
support Python < 2.7, these checks are useless.