Fix issue #6
The function icarus.util.apportionment calls the function heapq.nlargest
which throws an error if its first argument is not of int type. Such
an error is thrown even if the argument is an integer number of type
float (e.g., 2.0). This causes the apportionement function to throw an
error when using specific combinations of arguments.

This commit addresses this issue by casting the first argument of
heapq.nlargest to int.