Make single entry point for all commands
Up until now, to run a simulation with Icarus, a user had to
run the icarus.py file in the root of the project and to work
with result files use either mergeresults.py or printresults.py
in the scripts directory.

This commit consolidates all commands under a single entry point
whose source code is in icarus/main.py and, on installation is added
to the executable path as `icarus`.

As a result, a user can run simulations with `icarus run ...` and
process results with `icarus results ...`. This does not affect
using Icarus as a library.