Fix randomly failing tests in Python 3.5
It appears that using Python 3.5, there is no longer the guarantee that
Networkx's all_pairs_dijkstra_path will pick, in presence of equally
long path the one where the first diverging node is first in
lexicographic ordering. This causes some tests relying on this
assumption fail.

This commit addresses this by changing test cases that relied on this
assumption and ensuring that paths have different lengths.