Fix randomly failing TestHashroutingEdge test cases
Some test cases used to fail in Python 3.x because it appears there is
no longer the guarantee that Networkx's all_pairs_dijkstra_path will
pick, in presence of equally long paths, the one where the first
diverging node is first in lexicographic ordering. This caused some
tests relying on this assumption fail.

This commit fixes that.