Fix #21
The degree and betweenness centrality cache placement algorithms
incorrectly assumed that all nodes of a topology were cache candidates.
When this was not the case, cache nodes were still allocated cache capacity
proportionally to their degree or betweenness centrality but not all cache
budget was used.

This commit fixes corrects this problem and adds unit test cases covering
the degree and betweenness centrality cache placement algorithms.